+++ ts/features/messages/store/actions/index.ts
@@ -1,7 +1,43 @@
-import { ActionType, createAsyncAction } from "typesafe-actions";
+import { ActionType, createAction, createAsyncAction } from "typesafe-actions";
 import { ThirdPartyMessageWithContent } from "../../../../../definitions/backend/ThirdPartyMessageWithContent";
-import { UIMessageId } from "../../../../store/reducers/entities/messages/types";
 import { ServiceId } from "../../../../../definitions/backend/ServiceId";
+import { UIMessageId } from "../../../../store/reducers/entities/messages/types";
+import { MessageGetStatusFailurePhaseType } from "../../../../store/reducers/entities/messages/messageGetStatus";
+
+export type ThirdPartyMessageActions = ActionType<typeof loadThirdPartyMessage>;
+
+export type RequestGetMessageDataActionType = {
+  messageId: UIMessageId;
+  fromPushNotification: boolean;
+};
+
+export type SuccessGetMessageDataActionType = {
+  containsAttachments: boolean;
+  containsPayment?: boolean;
+  euCovidCerficateAuthCode?: string;
+  firstTimeOpening: boolean;
+  hasRemoteContent: boolean;
+  isPNMessage: boolean;
+  messageId: UIMessageId;
+  organizationName: string;
+  serviceId: ServiceId;
+  serviceName: string;
+};
+
+type FailureGetMessageDataActionType = {
+  blockedFromPushNotificationOpt?: boolean;
+  phase: MessageGetStatusFailurePhaseType;
+};
+
+export const getMessageDataAction = createAsyncAction(
+  "GET_MESSAGE_DATA_REQUEST",
+  "GET_MESSAGE_DATA_SUCCESS",
+  "GET_MESSAGE_DATA_FAILURE"
+)<
+  RequestGetMessageDataActionType,
+  SuccessGetMessageDataActionType,
+  FailureGetMessageDataActionType
+>();
 
 /**
  * The user requests the message third party content.
@@ -16,4 +52,10 @@ export const loadThirdPartyMessage = createAsyncAction(
   { id: UIMessageId; error: Error }
 >();
 
-export type ThirdPartyMessageActions = ActionType<typeof loadThirdPartyMessage>;
+export const resetGetMessageDataAction = createAction(
+  "GET_MESSAGE_DATA_RESET_REQUEST"
+);
+
+export const cancelGetMessageDataAction = createAction(
+  "GET_MESSAGE_DATA_CANCEL_REQUEST"
+);
