+++ ts/screens/messages/__tests__/MessageRouterScreen.test.tsx
@@ -5,7 +5,7 @@ import { appReducer } from "../../../store/reducers";
 import { UIMessageId } from "../../../store/reducers/entities/messages/types";
 import { renderScreenWithNavigationStoreContext } from "../../../utils/testWrapper";
 import { MessageRouterScreen } from "../MessageRouterScreen";
-import { getMessageDataAction } from "../../../features/messages/actions";
+import { getMessageDataAction } from "../../../features/messages/store/actions";
 import { ServiceId } from "../../../../definitions/backend/ServiceId";
 import * as ASD from "../../../store/hooks";
 import { MessageGetStatus } from "../../../store/reducers/entities/messages/messageGetStatus";
